% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absSSB}
\alias{absSSB}
\title{absolute SSB plot}
\format{An object of class \code{gg} (inherits from \code{ggplot}) of length 10.}
\usage{
absSSB
}
\description{
absolute SSB plot
}
\keyword{datasets}
