% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{bubblePlot}
\alias{bubblePlot}
\title{bubble plot of residuals}
\usage{
bubblePlot(rep, indices, oneStep = NULL, pch = 21)
}
\arguments{
\item{rep}{the report object of the fit}

\item{indices}{the indices data frame used in the data}

\item{oneStep}{optional oneStepPredict residuals}
}
\description{
Bubble plots of the residuals using either oneStepPredict residuals or not
}
