% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{datSetup}
\alias{datSetup}
\title{Setup data for SPAM model}
\usage{
datSetup(surveys, catch, landings, stock_wt, midy_wt, mat, M = 0.2,
  ages = NULL, years = NULL, plusGroup = NULL, idetect = 5e-04,
  cdetect = 5e-04, match3NOdims = TRUE, naz.rm = TRUE)
}
\arguments{
\item{surveys}{named list of surveys to include}

\item{catch}{data.frame of catch to include}

\item{landings}{data.frame of landings to include}

\item{stock_wt}{data.frame of stock weights}

\item{midy_wt}{data.frame of midyear weights}

\item{mat}{data.frame of maturity}

\item{ages}{optional vector of ages to include, if using plus group, keep ages above plusGroup in vector}

\item{years}{optional vector of years to include}

\item{plusGroup}{optional age to specify as a plus group, indices are summed by that age and weights taken to be the mean of ages in plus group}

\item{idetect}{The survey detection limit}

\item{cdetect}{The catch dectection limit}

\item{match3NOdims}{remove final row from C,C_zero,landings}

\item{naz.rm}{replaces NAs and zeros with dectection limits}

\item{recType}{recruitment option}
}
\description{
Constructs data and parameters to be usable by SPAM.
This accepts the data setup by mdata3Ps.R in the SPAM data-raw folder.
}
