% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldbsource.R
\name{lldbsource}
\alias{lldbsource}
\title{Debug TMB object code through LLDB}
\usage{
lldbsource(file, interactive = FALSE, object.name = NULL)
}
\arguments{
\item{file}{the file to debug}

\item{interactive}{whether to interactively control the lldb session.}

\item{object.name}{optional name of the compiled object if different from file name.}
}
\description{
This can be used to debug C++ from within R, in the same fashion as TMB's gdbsource,
except it replaces gdb with the lldb debugger. I wrote this as I could not get gdbsource to work on OS X 10.14.
}
