% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{logLik.spam}
\alias{logLik.spam}
\title{Returns the loglikelihood of a spam object}
\usage{
\method{logLik}{spam}(object, ...)
}
\arguments{
\item{object}{the spam object to get the loglikelihood of}
}
\description{
Returns the loglikelihood of a spam object
}
