% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mboundSSB}
\alias{mboundSSB}
\title{Bounds testings}
\format{An object of class \code{gg} (inherits from \code{ggplot}) of length 9.}
\usage{
mboundSSB
}
\description{
Bounds testings
}
\keyword{datasets}
