% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelOpts}
\alias{modelOpts}
\title{Result of running the model 100 times with every possible user configurable option}
\format{An object of class \code{data.frame} with 100 rows and 8 columns.}
\usage{
modelOpts
}
\description{
This is a table containing the AIC,BIC, time taken, likelihood, convergence message of
running SPAM 100 times on a specfic set parameters, this was done with every possible option
of correlation type for F and CRL along with the different recruitment curve options
}
\keyword{datasets}
