% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{oneStopPlot}
\alias{oneStopPlot}
\title{Plot a lot of different things}
\usage{
oneStopPlot(resObj, indName = NULL, facetAges = TRUE,
  residuals = FALSE, loess = FALSE, bubble = FALSE)
}
\arguments{
\item{resObj}{residuals spam object with extra junk for plotting}

\item{indName}{optional the name of the catch,survey,landing etc. to look at}

\item{facetAges}{facet wrap the plots by age? Takes sum of index and mean of residuals over ages if false}

\item{residuals}{plot the residuals instead of the line plots}

\item{loess}{add a loess smoother to the residuals?}

\item{bubble}{bubble plots of residuals?}

\item{oneStep}{optional oneStepPredict residuals to supply}
}
\description{
Compare the predicted to actual value of the survey or catch crl index by age or not.
Returns a ggplot plot. This works for both catch and surveys because everything is done in one
vector now.
}
