% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{pGroup}
\alias{pGroup}
\title{Helper function for generating plusGroup information}
\usage{
pGroup(x, plusGroup, sOrM = "survey", ages, years)
}
\description{
Helper function for generating plusGroup information
}
