% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{qPattern}
\alias{qPattern}
\title{Plot estimated q patterns for surveys}
\usage{
qPattern(sd.rep, indices, alpha = 0.05, qpname = "log_qparm")
}
\arguments{
\item{sd.rep}{the sd.rep object}

\item{indices}{the indices object}

\item{alpha}{alpha value for the confidence intervals}

\item{qpname}{name of q-parameters in report if different}
}
\description{
Function to plot the estimated q catchabilty patterns for surveys.
Returns a ggplot2 plot. Values are standardized to the max of the q.
}
