% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rDat}
\alias{rDat}
\title{Data readin and used}
\format{An object of class \code{list} of length 15.}
\usage{
rDat
}
\description{
Data required to run SPAM. Includes surveys, a list containing all surveys and catch for use with datSetup.
fstock_wt, which are FPCA smoothed backdated weights, fcomm_wt which are fpca smoothed commerical weights,
midy_wt which was the provided midy_wt file,mat which was the provided maturity file,stock_wt original provided stock wt,
DFO_stock_wt,catch data on it's own, geac and french survey indices,dfo survey indices
}
\keyword{datasets}
