% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{resid4plot}
\alias{resid4plot}
\title{Residuals by year, cohort, age and against expected values}
\usage{
resid4plot(dat)
}
\arguments{
\item{resObj}{the report object containing the residuals}
}
\description{
Returns a list containing all four ggplot objects which can then easily be combined with something like ggarrange.
}
