% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{residuals.spam}
\alias{residuals.spam}
\title{Returns the one step ahead residuals of a spam object and some other stuff for plotting}
\usage{
\method{residuals}{spam}(object, method = "oneStepGaussianOffMode", ...)
}
\arguments{
\item{object}{the spam object}

\item{method}{the one step ahead residuals to use}

\item{...}{options to give oneStepPredict}
}
\description{
Returns the one step ahead residuals of a spam object and some other stuff for plotting
}
