% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retL}
\alias{retL}
\title{Retrospective plots and retro stuff}
\format{An object of class \code{list} of length 5.}
\usage{
retL
}
\description{
Retrospective plots and retro stuff
}
\keyword{datasets}
