% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{spamFit}
\alias{spamFit}
\title{Fit the SPAM model}
\usage{
spamFit(data, parameters, indices, cVec, random = NULL, map = list(),
  control = list(), silent = TRUE, ...)
}
\arguments{
\item{dat}{list of data and parameters from spamSetup}
}
\description{
Basic wrapper so I get a class to make things like AIC eaiser.
}
