% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{spamSetup}
\alias{spamSetup}
\title{Setup data for SPAM model}
\usage{
spamSetup(surveys, catch, landings, keys, stock_wt, midy_wt, mat,
  recType = 0, corrFFlag = 0, M, years = NULL, ages = NULL,
  plusGroup = NULL)
}
\arguments{
\item{surveys}{named list of surveys to include}

\item{catch}{named list of catch to include}

\item{landings}{data.frame of landings to include}

\item{stock_wt}{data.frame of stock weights}

\item{midy_wt}{data.frame of midyear weights}

\item{mat}{data.frame of maturity}

\item{recType}{recruitment option}

\item{years}{optional vector of years to include}

\item{ages}{optional vector of ages to include, if using plus group, keep ages above plusGroup in vector}

\item{plusGroup}{optional age to specify as a plus group, indices are summed by that age and weights taken to be the weighted mean of ages in plus group by index}
}
\description{
Constructs data and parameters to be usable by SPAM.
Assumes data is in the format of ../data-raw/mdata3Ps.R with years.
}
