% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{spayPlot}
\alias{spayPlot}
\title{Spay plot of model fit}
\usage{
spayPlot(rep, indices)
}
\arguments{
\item{rep}{the report object containing the residuals}

\item{indices}{the indices data object containing the indices setup data}
}
\description{
Spay plot of model fit
}
