% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{surCompPlot}
\alias{surCompPlot}
\title{Survey/Catch composition plot}
\usage{
surCompPlot(rep, indices, indName = NULL, facetAges = TRUE,
  residuals = FALSE, oneStep = NULL, loess = FALSE)
}
\arguments{
\item{rep}{the report object containing the predicted indices}

\item{indices}{the indices list of info}

\item{indName}{optional the name of the catch,survey,landing etc. to look at}

\item{facetAges}{facet wrap the plots by age? Takes sum of index and mean of residuals over ages if false}

\item{residuals}{plot the residuals instead of the line plots}

\item{oneStep}{optional oneStepPredict residuals to supply}

\item{add}{a loess smoother to the residuals?}
}
\description{
Compare the predicted to actual value of the survey or catch index by age or not.
Returns a ggplot plot. This works for both catch and surveys because everything is done in one
vector now.
}
