% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topMod}
\alias{topMod}
\title{The top chosen model, residuals, data, etc.}
\format{An object of class \code{list} of length 3.}
\usage{
topMod
}
\description{
The top chosen model, residuals, data, etc.
}
\keyword{datasets}
