% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{tsCIplot}
\alias{tsCIplot}
\title{Plot line graph with confidence interval from sd.report object}
\usage{
tsCIplot(sName, sd.rep, years, alpha = 0.05, exp = FALSE,
  add = FALSE, color = "black", fcolor = "grey70")
}
\arguments{
\item{sName}{the name of the series to plot}

\item{sd.rep}{the sd.report object}

\item{years}{vector of corresponding years}

\item{alpha}{alpha value for confidence intervals}

\item{exp}{take the exponent of the value?}

\item{add}{return a list of gproto objects to add together plots?}

\item{color}{the color of the line}

\item{fcolor}{the color of the confidence interval}
}
\description{
Given a name of something in an sd.report object and the number of corresponding
years, plot the time series and the corresponding confidence intervals and return a ggplot object
}
