% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCplots.R
\name{resid4plot}
\alias{resid4plot}
\title{Residuals by year, cohort, age and against expected values}
\usage{
resid4plot(rep, indices, oneStep = NULL)
}
\arguments{
\item{rep}{the report object containing the residuals}

\item{indices}{the indices data frame}

\item{oneStep}{optional oneStepPredict residuals to use instead}
}
\description{
Returns a list containing all four ggplot objects which can then easily be combined with something like ggarrange.
}
